ARM: dts: msm: sun: Add SOCCP_SOCCP_SPARE_REG0 to check SOCCP status
TCSR register can't be relied on to check D0 status as it is updated
by HW and SW on the SOCCP might not be in proper state. In that case,
if APSS tries to communicate with SOCCP, we see D3 to D0 transition
failures. To avoid such issues, read SOCCP_SOCCP_SPARE_REG0 to check
SOCCP D0 status and stick to TCSR for D3 status check.

Change-Id: Icee37cddb0b7ef303962cab0d9a8f37a211a05da
Signed-off-by: Gokul krishna Krishnakumar <quic_gokukris@quicinc.com>