ARM: dts: msm: Increase virtio-mem maximum size
ZRAM is being disabled on tuivm due to observed stability issues due to
ZRAM's use of GFP_MEMALLOC. Increase maximum movable zone size to
compensate.

When kswapd swaps out pages, PF_MEMALLOC is set, which results in zram
ignoring the min_watermark, and lowmem_reserve. This is problematic
because we intend for movable memory to stay in ZONE_MOVABLE, rather
than falling back to ZONE_NORMAL. If the later occurs to such a high
degree that all ZONE_NORMAL memory is consumed, as is observed to be
the case during a "C2PA" testcase, attempts to plug in additional
ZONE_MOVABLE memory may fail due to requiring a small amount of
GFP_KERNEL memory.

Change-Id: Ic847e90514f80ccb91daa98849440d7a9b97f156
Signed-off-by: Patrick Daly <quic_pdaly@quicinc.com>