dt-bindings: soc: qcom: Snapshot of PMIC glink binding
The PMIC GLINK service, running on a coprocessor on some modern Qualcomm
platforms and implement USB Type-C handling and battery management.
This binding describes the component in the OS used to communicate with
the firmware and connect it's resources to those described in the
Devicetree, particularly the USB Type-C controllers relationship with
USB and DisplayPort components.

This is a snapshot taken as of common kernel branch
commit 254c101efde7 ("dt-bindings: soc: qcom: pmic-glink: allow
orientation-gpios").

Change-Id: I2afdd684dbaede3e0cedca29e684620e491ce892
Signed-off-by: Anjelique Melendez <quic_amelende@quicinc.com>