ARM: dts: msm: Remove clocks property form CPU nodes
The initialization of the OPP table from the device tree is failing
because the OPP framework expects the CPUFreq node to act as a clock
provider due to the presence of the clocks property in CPU nodes.
However, the qcom-cpufreq-hw scaling driver doesn't have the support
for handling the CPUFreq node as a clock provider, resulting in an
-EPROBE_DEFER error. Thus, to resolve this issue, remove the clocks
property from the CPU nodes.

Change-Id: I243807f58dc82c55f4ec390c09752b8652ac2706
Signed-off-by: Imran Shaik <quic_imrashai@quicinc.com>