ARM: dts: msm: Fix IOMMU address reserve definition
In order to ensure that IOMMU returns a 32 bit address, extend the
address and size cells to 2, in order to define a region that will
block 0xf0000000--0xffffffffffffffff.  By increasing the address and
size cells for the parent node, the DWC3 child node has to also
change the 'reg' entry to accommodate the added values.

Use upstream compatible DT property "iommu-addresses" instead
of "qcom,iommu-dma-addr-pool" since its deprecated.

Change-Id: Ie75c33ff86b33aeede469f67038eaae881c4dc3a
Signed-off-by: Mohana LNU <quic_mohalnu@quicinc.com>