ARM: dts: msm: Correct reg fields of QUPv3, GPI and i2chub dt nodes
As part of commit '9bc13fb3dc10a04fe643b554f348a022ee917695
("ARM: dts: msm: Add QUPv3 and GPI DT nodes on Canoe")',
reg fields of QUPv3, GPI and i2chub dt nodes,as well as
the qcom,iommu-geometry property of QUPv3, were not
aligned according to the SoC address and size cells.

So, aligned the reg fields of the QUPv3, GPI, and i2chub
DT nodes, as well as the qcom,iommu-geometry property of
QUPv3, according to the SoC address and size cells.

Change-Id: I26ab7c43266bc22ea949e2381a3711ac2aeb9dff
Signed-off-by: Krishna Chaithanya Reddy G <quic_kgangapu@quicinc.com>