ARM: dts: msm: Add wcd_usbss reference to USB on canoe
The USB D+/D- signal lines are first routed through the WCD939x USB
subsystem before connecting to the USB controller on MTP and some QRD
based platforms for canoe. Add a phandle to the former to the USB
device node on such platforms. This will allow the USB driver to
control the D+/D- switches when USB 2.0 functionality is needed.

Change-Id: Ic5c9d3b319f766bf61c35bae64dde99329dc9eae
Signed-off-by: Elson Roy Serrao <quic_eserrao@quicinc.com>