ARM: dts: msm: Add NAND/SD card/eMMC support for sdxkova RCM platform
Enable NAND and SD card for RCM MBB platform.
Enable eMMC for RCM MBB V3 platform and disable nand for the same
platform as that is eMMC based.
To support this remove SD card properties using the /delete-node/
directive and reusing the same sdhc node for eMMC for RCM MBB V3.

Change-Id: I1f3b160c982c0bf7682885a0c2e0a419384f5ca4
Signed-off-by: Sarthak Garg <quic_sartgarg@quicinc.com>