ARM: dts: msm: Temporarily Re-enable ZONE_DMA32 on canoe
Temporarily re-enable ZONE_DMA32 on KP5.0 for canoe to allow LTS
6.12 merge to move forward while debugging an issue related to memory
allocation for a driver using smmu bypass for device memory allocations.

Once the issue is root caused and fixed, a follow up patch will be pushed
to disable ZONE_DMA32 once more.

Note: There should not be any functional change to behavior on Kaanapali
so far because LTS 6.12 is what was supposed to bring changes that will
effectively disable ZONE_DMA32.

Change-Id: I60275dbe946c08cdd6b1f3141b78219e27c2e110
Signed-off-by: Oreoluwa Babatunde <quic_obabatun@quicinc.com>