ARM: dts: msm: Add Super-Speed USB support on Alor
Add QMP phy node in usb file to support super-speed usb usecases.
Change includes all the necessary resources along with the init
sequence of the phy required to execute.
Also assigning the usb phys to the controller and moving nop phy to usb
files for ease of access for debugging.

Change-Id: I987e1fe2b4f3da005489a6559621883018dfb48e
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>