ARM: dts: msm: Fix BT inquiry issue by updating UART Rx wakeup in SM6150
BT inquiry stop was not functioning correctly. The system entered
sleep state, and the Rx wakeup interrupt failed to trigger,
resulting in missing Rx packets on the Bluetooth side.

Updated the 4uart node pinctrl configuration for Rx wakeup to
ensure proper interrupt handling and resolve the BT inquiry issue
on the SM6150 platform.

Change-Id: I2ad71153d84ed822af0ce4ed6fa8bad24708a7cb
Signed-off-by: Nandi Bhoopathi <quic_nandibho@quicinc.com>
Signed-off-by: Asit Shah <quic_asitshah@quicinc.com>