ARM: dts: msm: Configure usbcontroller for yupik
On legacy soc like yupik-iot where primary and
secondary usb available, adbd is started with
secondary usb because ls -1 /sys/class/udc lists
in a to z order in busybox.
Due to this adbd is not working in kbdev. Hence configure
androidboot.usbcontroller with primary usb controller which
will be used for adb.

Change-Id: Iec761e433744df3f1eee0a6fa40adf3e7130732e
Signed-off-by: Saranya R <quic_sarar@quicinc.com>
Signed-off-by: Swetha Chikkaboraiah <quic_schikk@quicinc.com>