ARM: dts: msm: Add VDAC support for PMAR2230
Add support for new QCOM VDAC driver in PMAR2230 device tree, to enable
support for generating analog control signals for applications such as
opacity control in AR headsets.

Also add 2 new VADC channels ATEST1/2, that can be used to readback the
DAC output voltage and ground reference voltage to check if the VDAC is
working as expected and add new NVMEM cell for MISC peripheral, which is
needed to internally route the DAC output to ADC channels inside PMAR2230,
to support this readback via ADC channels.

Change-Id: I13fd09e5ce00cf1f4d296cb19aca8953f4fb9295
Signed-off-by: Brindha T <brint@qti.qualcomm.com>
Signed-off-by: Dhruvin Rajpura <drajpura@qti.qualcomm.com>