ARM: dts: msm: Use upstream PM8008 drivers for yupik IOT target
Change the DT file as required by the upstream PM8008
drivers - mfd/qcom-pm8008 and regulator/qcom-pm8008-regulator.

Also modify the min/max voltage levels for a few LDOs since the
regulator registrations were failing when the previously set voltage
level by the bootloader was not matching with the min/max levels
set in the DT. Clients would need to ensure that they continue
to use the same min/max values as were previously being used by them
with the downstream driver on 5.4 kernel.

Change-Id: If11bd5e24e0cef9b5bd65261cf20fb3532f731dd
Signed-off-by: Shilpa Suresh <c_sbsure@qti.qualcomm.com>