ARM: dts: qcom: Disable halt feature at PCIe for sdxpinn
Change is to disable halt feature for sdxpinn at PCIe level, as
NOC already supports ordering feature. Enabling this at PCIe level
degrade the performance when bandwidth is above 9 Gbps.

This change is only enabled for WKK, as for HMT and ETH we dont
anticipate to cross the 9Gpbs bottleneck for these attaches.

Propagated from sdxpinn change 5596777.

Change-Id: I19c7fa729b46e10539e0b65407ea5c1ea93dd7dd
Signed-off-by: Ashishkumar Vijayakumar <ashivija@qti.qualcomm.com>