ARM: dts: msm: Add u3 susphy quirk to avoid susphy being setfor canoe
Keeping susphy enabled lets the phy to suspend when controller is in
idle state, however in scenarios where host to device transition is
involved this could cause problems since in transition if the phy
suspends the dwc controller will loose it's valid pipeclk and would try
to access invalid addresses.

Fix this by adding the quirk which would disable this feature.

Change-Id: Icf349487020031c972f58e0e4d386ced76221208
Signed-off-by: Akhila Chevuru <achevuru@qti.qualcomm.com>