ARM: dts: qcom: Fix C610 crash due to IOMMU misconfiguration on Slimbus
On the C610 platform, a crash was observed during Slimbus audio code
operations due to incorrect IOMMU configuration. The Slimbus
peripheral was not properly mapped, leading to memory access
violations and system instability.

This patch corrects the IOMMU settings in the device tree to ensure
proper mapping for Slimbus, resolving the crash and restoring stable
functionality on the C610 platform.

Change-Id: I90b181d7c937c3ec30c62145c2fd391ae54f5835
Signed-off-by: Nandi Bhoopathi Rao <quic_nandibho@quicinc.com>