ARM: dts: msm: Enable emergency shutdown with threshold at 2.7V
It's possible that even at 0% device may fail to shutdown if the
android power service is not working properly. In such a scenario,
device will continue to draw power after reaching 0%, this can lead
to deep discharge of battery, to the point that it may become
unrecoverable to boot with it, even with a charger connected or it
may take lot of time to charge it to normal levels.

To avoid this problem, set the qcom,shutdown-voltage to 2.7v, this
will ensure that the device will go for shutdown if the soc = 0%
and voltage falls below 2.7v. This will ensure that even if the
android fails to trigger shutdown, the charger driver will trigger
an emergency shutdown to save the battery from getting deeply
discharged.

Change-Id: Ie30f835f860b733378079cb667d9d3745fdaedb4
Signed-off-by: Rakesh Kota <rakesh.kota@oss.qualcomm.com>
Signed-off-by: Ayyagari Ushasreevalli <quic_aushasre@quicinc.com>