ARM: dts: msm: Increase pipe clock toggles during L1SS entry for Alor
Increase the number of pipe clock toggles that will occur after
phystatus goes high at the output of the PHY during L1SS/P2 entry
in PCS_PCIE_POWER_STATE_CONFIG6 register, so that if the phy and
controller goes out of sync this may help us. The number of pipe
clock toggles is equal to (4*value)+1.

Change-Id: I6ec3f66adb1fc8cc30fbe21d7c7e98a4cd7f3f4c
Signed-off-by: Vivek Pernamitta <quic_vpernami@quicinc.com>