ARM: dts: msm: Vote MMNOC BW for gcc_camss_top_gdsc on Vienna
The gcc_camss_top_gdsc causes camera to permanently assert qactive
when it's enabled. If MMNOC attempts to collapse when the GDSC is
enabled, then the qactive signal will cause BCM to get stuck in the
collapse sequence. This will eventually lead to rpmh driver timeouts
when requested for MMNOC. Fix this by explicitly voting for MMNOC on
behalf of the GDSC.

Change-Id: I9c8f58dbc6568be55e9cbb35e0c2297c570d7958
Signed-off-by: Vishnu Mani Vardhan Gadepalli <vgadepal@qti.qualcomm.com>